adding some stuff

+++ helpful/helpful.cpp
@@ -25,10 +25,10 @@ string sub_binary(string input_1, string input_2) {
     return output;
 }
 
-int binary_to_int(string input) {
+unsigned long binary_to_int(string input) {
     return bitset<32>(input).to_ulong();
 }
 
-string int_to_binary(int input) {
+string int_to_binary(unsigned long input) {
     return bitset<32>(input).to_string();
 }
